requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# Vision-Language Models
transformers>=4.30.0
timm>=0.9.0
einops>=0.6.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.5.0
opencv-python>=4.7.0
scikit-learn>=1.2.0
scikit-image>=0.20.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Evaluation Metrics
nltk>=3.8
rouge-score>=0.1.2
bert-score>=0.3.13

# Logging and Experiment Tracking
tensorboard>=2.13.0
wandb>=0.15.0
tqdm>=4.65.0

# Jupyter and Development
jupyter>=1.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
albumentations>=1.3.0
